Pseudo guard clause for donut chart
Hack together a pseudo guard clause for making sure the donut chart has finished rendering before we take a screenshot.

Note that we use actual guard clauses, but rely on knowing what height & width _should be_ to indicate it is nearly finished (for now).

The tricky part that is not currently solved is that there is still a tiny bit of animation left at the time the parent reaches max width (filling in wedge gaps).

This is clearly fragile and won't hold forever, but it is lightning fast compared to arbitrary waits.