Components don't call firestore init directly
The components no longer call the method for getting a reference to firestore. They still know quite a bit about firestore (especially those subscribing), but now it should be possible to mofiy or adjust what we would pass back to them in one fell swoop, such as we might want to do if pointing at an emulator for tests.

Beyond that, we are on the way to some testability benefits to the individual components, since we have a more malleable seam where could potentially inject mocks (with a bit more effort).

Also remove some unused data props named `db`, since they showed up when searching for 'db' references.