All components use same firebase reference
The components now reference firebase as an instance variable.

This prevents an issue that was uncovered when we started using the emulator, whereby repeated configurations/setup calls caused firebase to get grumpy with us. This was seemingly never an issue for the cloud hosted version, or perhaps we just weren't toggling any settings at that point.

Note that we are hard-coding to use the emulator at the moment (that is the motivating use-case for these changes), but we will follow up on that shortly to make it toggle-able.