Support toggling firebase emulator use via env var
Support toggling the emulator via an environment variable that is configured to be toggled by the build _mode_ (see: https://cli.vuejs.org/guide/mode-and-env.html#modes). Setting the 'e2e' mode causes the vars from `.env.e2e` to be loaded, which then set the needed environment variables.

Given the way these files work (allow overrides, including local, as the project already uses), it should be quite easy to use the emulator for all local development, if one so chooses (and has the emulator set up - documentation forthcoming)

Also, add a new script, "serve:e2e", that will start the app with this mode set.

Note that we also remove the explicit `useEmulator` flag in favor of the newly-added mode/variable toggle.