Allow Firebase to auto-detect long-polling
Adding an experimental setting for Firebase SDK to auto-detect long-polling.

This would seem to be needed due to Cypress, which is effectively throttling requests to Firebase, compared to what it would be expecting. See answer here: https://stackoverflow.com/a/61619639

It appears as well that there is an open issue for this on Cypress (2347)