Adding firestore emulator data for ProjectOne test
Initial add of firestore emulator meta as well as data for ProjectOne test.

This can be used together with `--import` arg when starting the emulator to pre-populate the exported data as seed data for our tests.

Note that we have a very simple test (that is also quite short) at the moment, so something like resets between tests are not yet needed, though they can be a very good idea, and could some day make sense.