Add e2e test for ProjectTwo including seed data
Adding an e2e test for ProjectTwo along with the seed data for the firestore emulator.

Note that compared to the first test, this one creates data that is not so easy to reverse, so reseting this test involves either restarting the emulator or removing the records by hand.

That makes debugging this test not a lot of fun. So, tests like this one are good motivations for adding state reset between the tests/runs.