Adding e2e test for ProjectThree
Adding a simple e2e test for ProjectThree along with the seed data.

Similar to the test for ProjectTwo, this one modifies state, and it isn't possible to revert it from within the app currently. So, at some point it would be good to add a rollback/reset before re-running (as for debug and similar)