Move extended timeout to initial dynamic elements
Make sure that the e2e tests apply to extended timeout to the first _dynamic_ elements (created by d3, and not part of the initial template).

This we do to make sure that we only proceed with additional verifications once we know that we've started rendering 'updates' that include the data from firestore.