Run e2e on Windows for now
Currently, it is not possible to import emulator exports run on Windows into an emulator running on linux.

Our current canonical import file was exported on Windows, and I cannot get the emulator to run in WSL for me _at all_ (with or without import). Unfortunately debugging experience is catastrophically bad, and all I see is a generic "EACCESS" error, with no further details.

So, go ahead with a Windows runner for now, because I have no desire to waste more time on debugging the half-baked emulator.