Run e2e on Windows for now
Currently, it is not possible to import emulator exports run on Windows into an emulator running on linux.

Our current canonical import file was exported on Windows. Previous attempts to get the emulator to run in WSL failed (with or without import - generic "EACCESS" error, with no further details).

So, go ahead with a Windows runner for now, because it is valuable enough to have these tests run in the CI.