Update Sass Loader to v10
Had an issue running the project in a clean install situation (trying to run in WSL), and it seems somehow an update to Sass Loader got missed during recent updates.

Note that v10 seems to be as far as we can go right now. V11 and v12 would cause issues when trying to run locally (in debug).