Move Preview away from Bootstrap-Vue
Move the preview component to use only Vue & Bootstrap, and not Bootstrap-Vue.

Note that to keep the carousel working, I seemed to have to import jQuery, unfortunately.

Also, animation of the carousel works a little different now, and seems to only start animating once one of the arrows have been clicked (despite using the "data-ride" attribute, which is supposed to solve this.