Add size-limit for bundle size tracking
It covers the same need as Bundlewatch, but is more active, and has a few more features.

It also has some neat features like calculating load time and defining limits in terms of time, as well as options to support adding webpack plugins, and webpack-bundle-analyzer style output (via `--why`).

To start with, I have set the limits to the baseline  + 25%, which is suggested in the documentation.

Note as well that bundle sizes _appear_ different than those from Bundlewatch, but that is merely a kB vs. KiB thing. It seems size-limit shows in kB by default. Note that it is possible to define the limits in KiB, if we so choose.