Upgrade test packages for Vue 3
Upgrade @vue/test-utils to version 'next'.

Also install vue-jest (as vue3-jest), since it seems to be something of an unnamed/missing peer. (We were already using it to transform compnents for test, but hadn't explicitly imported it before).

Also make some necessary changes to the tests for updated vute-test API.

Last but not least, we now seem to need to set the environment to jsdom (else we don't have browser globals like `document`)