Switch to LTS tag for image
Use the LTS tag (an alias) to always get the most recent LTS-channel version of Node that Circle CI knows about.

In part this is just a good idea.

But also, a week after the release of v18.13 they still don't seem to have a matching image, so the 18.13.0-browsers tag fails at the moment.