Add Vite packages
Based on Step #1 of the guide here: https://vueschool.io/articles/vuejs-tutorials/how-to-migrate-from-vue-cli-
to-vite/

Note that the plugins we need to use here have changed since the guide was written (along with the package versions).

Most notably, we no longer need the package "vite-plugin-vue2", as it only supports Vue 2.6 or earlier.

Instead, we use @vitejs/plugin-vue2, which supports use of SFCs with Vite. It is also used instead of the @vit
ejs/vue-plugin for now (until we upgrade to Vue 3).