Replace CLI config with Vite config
Still following along with the same guide as before (Step #3 - https://vueschool.io/articles/vuejs-tutorials/how-to-migrate-from-vue-cli-to-vite/), but the alias setting is from a more recently generated Vite project.

Note though that I've chosen to add the main Vite config as a TS file, which, while we don't have explicit TS support, seems to be accepted, and gets us a tiny bit more type safety.