Move index.html to project root and update to be vite-compatible
Vite expects it to be at project root, as it uses it as a root entry point.

Further explanation: https://vitejs.dev/guide/#index-html-and-project-root

Along these same lines, import the script directly, and do so as a module (key Vite feature!)

Last but not least, remove Webpack-specific placeholder.

(Step 4 in the guide we've been following: https://vueschool.io/articles/vuejs-tutorials/how-to-migrate-from-vue-cli-to-vite/ )