Update main development scripts to use Vite
Update the main dev scripts, including lint, but not (yet) test, to use Vite, or ESLint directly in the case of lint.

Note that I'm including TS files in the pattern for lint as a proactive measure for when a TS migration comes along, but there are not currently any TS files in the project. (Easy to forget these things later)

More or less Step #5 from: https://vueschool.io/articles/vuejs-tutorials/how-to-migrate-from-vue-cli-to-vite/