Update environment variables
Update the environment variables to match vite convention.

This includes both the default variables (such as for determining if PROD), but also those we've defined for Firebase config.

This is Step #6 from: https://vueschool.io/articles/vuejs-tutorials/how-to-migrate-from-vue-cli-to-vite/

For more details on environments and modes in Vite, see here: https://vitejs.dev/guide/env-and-mode.html

ESLint config is also updated to allow use of import.meta (needs to be a module).