Correct missing PostCSS plugin error
An empty PostCSS config file (added on project init) was causing something in the Vitest pipeline to look for a PostCSS plugin, which caused an error when trying to run the tests.

As we aren't yet using PostCSS, we can simply remove the config for now, which instantly solved the issue.