Update Size Limit for ESBuild bundler
Swap out Size Limit Webpack packages for those appropriate for an app with ESBuild as its bundler, which applies to us after the migration to Vite.

Note that the previous paradigm for measuring chunk size doesn't entirely apply any more.

One major difference is that we have no separate chunks for vendor CSS (didn't see a way to make this possible with basic settings).

Another difference is that we have to use a plugin for vendor chunking of JS, as Vite no longer uses this strategy by default as of 2.8 - see here: https://vitejs.dev/guide/build.html#chunking-strategy

Lastly, even with Vendor chunking, we seem to see a very high difference in observed bundle size by Size Limit for our app JS file once calculations are run. This is almost certainly due to the inclusion of dependencies in Size Limit's calculation, and probably also influenced by the loading behavior of ES Modules, which we're now using (For more info on that, see here: https://hacks.mozilla.org/2018/03/es-modules-a-cartoon-deep-dive/ )