Update core packages to Vue 3-compatible versions
Update the core Vue and toolchain packages to versions that are compatible with Vue 3.

Note that this required a `--force` flag to override dependencies (helps us break this into incremental steps).

Also included is an update to v5 Vite, which, since we are overriding anyway, helps us skip a bunch of probably unnecessary intermediate steps.

Not following any one source 100%, but this combo of updates is loosely based on information from these sources:
- Migration build (note that many of the versions listed here are quite dated by now): https://v3-migration.vuejs.org/migration-build.html
- Release notes from the newest Vue, v3.4: https://blog.vuejs.org/posts/vue-3-4

Note that a few previously deprecated features were finally dropped in v3.4. They are not in use in our project, so an update directly to v3.4 wasn't problematic for us.