Update to Vue Router v4
Update to latest version of Vue Router, so that we are compatible with Vue 3.

Note that `--force` was also required for this step as well.

This required only a minimal amount of changes to router setup for changed APIs, due to the project's small size.

Beyond that, we switch to the new way of instantiating the app (in Vue 3), since we would need to do that anyway at some point, and the router wouldn't mount with the old method.

Migration guide available here: https://router.vuejs.org/guide/migration/