Update vitest to v1
Update vitest to v1, and perform the switch to v8 coverage provider (used since 0.32, released in early June 2023).

Process:
- Remove c8 coverage provider
- Install new package versions
- Uninstall c8 (possibly could be done sooner, was just initially forgotten)

While it is generally a good idea to keep our test runner up-to-date with the other packages (especially vite), we are trying to track down an error possibly caused by conflicting esbuild versions, in part coming from vitest packages.

With these changes, we are down to only two esbuild versions in our resolved dependencies (from three previously).

The error we were seeing: 'Error: The package "@esbuild/linux-x64" could not be found, and is needed by esbuild.'