Update size-limit packages
Update to the newest major version of size-limit packages, which gets us down to a single resolved version of esbuild.

Also bump the limit for the bundle, since we seem to have increased bundle size with the updates (compat package?)

Note that the bundle check is rather broken at the moment anyway (also counts vendor packages - see comment in config).