Remove unused prop
A previous comment indicated that this was used for testing while Bootstrap Vue was still in use.

The tests run fine without this prop, as does the app, so it doesn't seem to be needed any more.

This also removes a warning about attribute coercion from the Vue 3 compat build.